+++ parse.cc
@@ -188,8 +188,8 @@ parse_coordinates(const char* str, int datum, const grid_type grid,
                 &lathemi, &deg_lat, &lat, &lonhemi, &deg_lon, &lon, &result);
     valid = (ct == 6);
     if (valid) {
-      lat = (double)deg_lat + (lat / (double)60);
-      lon = (double)deg_lon + (lon / (double)60);
+      lat = (double)deg_lat + (lat / 60.0);
+      lon = (double)deg_lon + (lon / 60.0);
     }
     break;
 
@@ -200,8 +200,8 @@ parse_coordinates(const char* str, int datum, const grid_type grid,
                 &result);
     valid = (ct == 8);
     if (valid) {
-      lat = (double)deg_lat + ((double)min_lat / (double)60) + (lat / (double)3600.0);
-      lon = (double)deg_lon + ((double)min_lon / (double)60) + (lon / (double)3600.0);
+      lat = (double)deg_lat + ((double)min_lat / 60.0) + (lat / 3600.0);
+      lon = (double)deg_lon + ((double)min_lon / 60.0) + (lon / 3600.0);
     }
     break;
 
@@ -266,7 +266,7 @@ parse_coordinates(const char* str, int datum, const grid_type grid,
 
   if (datum != DATUM_WGS84) {
     double alt;
-    GPS_Math_Known_Datum_To_WGS84_M(lat, lon, (double) 0.0,
+    GPS_Math_Known_Datum_To_WGS84_M(lat, lon, 0.0,
                                     &lat, &lon, &alt, datum);
   }
